feat: add support for yoshi-ruby mono repo
