+++ src/releasers/node.ts
@@ -1,0 +1,87 @@
+import { ReleasePR, ReleaseCandidate } from '../release-pr';
+
+import { ConventionalCommits } from '../conventional-commits';
+import { GitHubTag } from '../github';
+import { checkpoint, CheckpointType } from '../util/checkpoint';
+import { Update } from '../updaters/update';
+import { Commit } from '../graphql-to-commits';
+
+// Generic
+import { Changelog } from '../updaters/changelog';
+// JavaScript
+import { PackageJson } from '../updaters/package-json';
+import { SamplesPackageJson } from '../updaters/samples-package-json';
+
+export class Node extends ReleasePR {
+  private async nodeRelease() {
+    const latestTag: GitHubTag | undefined = await this.gh.latestTag();
+    const commits: Commit[] = await this.commits(
+      latestTag ? latestTag.sha : undefined
+    );
+
+    const cc = new ConventionalCommits({
+      commits,
+      githubRepoUrl: this.repoUrl,
+      bumpMinorPreMajor: this.bumpMinorPreMajor,
+    });
+    const candidate: ReleaseCandidate = await this.coerceReleaseCandidate(
+      cc,
+      latestTag
+    );
+
+    const changelogEntry: string = await cc.generateChangelogEntry({
+      version: candidate.version,
+      currentTag: `v${candidate.version}`,
+      previousTag: candidate.previousTag,
+    });
+
+    // don't create a release candidate until user facing changes
+    // (fix, feat, BREAKING CHANGE) have been made; a CHANGELOG that's
+    // one line is a good indicator that there were no interesting commits.
+    if (this.changelogEmpty(changelogEntry)) {
+      checkpoint(
+        `no user facing commits found since ${
+          latestTag ? latestTag.sha : 'beginning of time'
+        }`,
+        CheckpointType.Failure
+      );
+      return;
+    }
+
+    const updates: Update[] = [];
+
+    updates.push(
+      new Changelog({
+        path: 'CHANGELOG.md',
+        changelogEntry,
+        version: candidate.version,
+        packageName: this.packageName,
+      })
+    );
+
+    updates.push(
+      new PackageJson({
+        path: 'package.json',
+        changelogEntry,
+        version: candidate.version,
+        packageName: this.packageName,
+      })
+    );
+
+    updates.push(
+      new SamplesPackageJson({
+        path: 'samples/package.json',
+        changelogEntry,
+        version: candidate.version,
+        packageName: this.packageName,
+      })
+    );
+
+    await this.openPR(
+      commits[0].sha,
+      `${changelogEntry}\n---\n`,
+      updates,
+      candidate.version
+    );
+  }
+}
