+++ test/updaters/fixtures/CHANGELOG-ruby.md
@@ -1,0 +1,46 @@
+# Release History
+
+### 0.5.0 / 2019-07-08
+
+* Support overriding service host and port
+
+### 0.4.2 / 2019-06-11
+
+* Add VERSION constant
+
+### 0.4.1 / 2019-04-29
+
+* Add AUTHENTICATION.md guide.
+* Extract gRPC header values from request.
+
+### 0.4.0 / 2019-03-11
+
+* Add v1beta1 API version
+
+### 0.3.0 / 2018-12-10
+
+* Add support for Regional Clusters.
+  * Client methods deprecate many positional arguments in
+    favor of name/parent named argument.
+  * Maintains backwards compatibility.
+
+### 0.2.2 / 2018-09-20
+
+* Update documentation.
+  * Change documentation URL to googleapis GitHub org.
+
+### 0.2.1 / 2018-09-10
+
+* Update documentation.
+
+### 0.2.0 / 2018-08-21
+
+* Move Credentials location:
+  * Add Google::Cloud::Container::V1::Credentials
+  * Remove Google::Cloud::Container::Credentials
+* Update dependencies.
+* Update documentation.
+
+### 0.1.0 / 2017-12-26
+
+* Initial release
